Creating appointmentsController
