+++ src/modules/appoitments/infra/http/controllers/AppointmentsController.ts
@@ -1,0 +1,20 @@
+import { Request, Response } from 'express';
+import { parseISO } from 'date-fns';
+import { container } from 'tsyringe';
+import CreateAppointmentService from '@modules/appoitments/services/CreateAppointmentService';
+
+export default class AppointmentsController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { user_id, date } = request.body;
+
+    const parsedDate = parseISO(date);
+    const createAppointment = container.resolve(CreateAppointmentService);
+
+    const appointment = await createAppointment.execute({
+      user_id,
+      date: parsedDate,
+    });
+
+    return response.json({ appointment });
+  }
+}
